chore: move `adv.push_smtpeek` to stdlib event
