+++ stdlib/tests/collections/smt.rs
@@ -1,3 +1,5 @@
+use miden_stdlib::EVENT_SMT_PEEK;
+
 use super::*;
 
 // TEST DATA
@@ -143,7 +145,8 @@ fn test_smt_set_empty_value_to_empty_leaf() {
 fn test_set_advice_map_empty_key() {
     let mut smt = Smt::new();
 
-    let source = "
+    let source = format!(
+        "
     use.std::collections::smt
     # Stack: [V, K, R]
     begin
@@ -160,7 +163,7 @@ fn test_set_advice_map_empty_key() {
         # => [K, R_new, V]
 
         # Fetch what was stored on advice map and clean stack
-        adv.push_smtpeek dropw dropw
+        emit.{EVENT_SMT_PEEK} dropw dropw
         # => [V]
 
         # Push advice map values on stack
@@ -171,7 +174,8 @@ fn test_set_advice_map_empty_key() {
         assert_eqw
         # => [K]
     end
-    ";
+    "
+    );
 
     let key = RpoDigest::new([41_u32.into(), 42_u32.into(), 43_u32.into(), 44_u32.into()]);
     let value: [Felt; 4] = [42323_u32.into(); 4];
@@ -186,7 +190,8 @@ fn test_set_advice_map_empty_key() {
 fn test_set_advice_map_single_key() {
     let mut smt = Smt::with_entries(LEAVES).unwrap();
 
-    let source = "
+    let source = format!(
+        "
     use.std::collections::smt
     # Stack: [V, K, R]
     begin
@@ -203,7 +208,7 @@ fn test_set_advice_map_single_key() {
         # => [K, R_new, V]
 
         # Fetch what was stored on advice map and clean stack
-        adv.push_smtpeek dropw dropw
+        emit.{EVENT_SMT_PEEK} dropw dropw
         # => [V]
 
         # Push advice map values on stack
@@ -214,7 +219,8 @@ fn test_set_advice_map_single_key() {
         assert_eqw
         # => [K]
     end
-    ";
+    "
+    );
 
     let key = LEAVES[0].0;
     let value: [Felt; 4] = [42323_u32.into(); 4];
